+++ packages/block-editor/src/components/off-canvas-editor/link-ui.js
@@ -3,6 +3,7 @@
 /**
  * WordPress dependencies
  */
+import { __unstableStripHTML as stripHTML } from '@wordpress/dom';
 import { Popover, Button } from '@wordpress/components';
 import { __ } from '@wordpress/i18n';
 import { switchToBlockType } from '@wordpress/blocks';
@@ -80,7 +81,7 @@ function LinkControlTransforms( { clientId } ) {
 	];
 
 	const transforms = blockTransforms.filter( ( item ) => {
-		return featuredBlocks.includes( item?.name );
+		return featuredBlocks.includes( item.name );
 	} );
 
 	if ( ! transforms?.length ) {
@@ -123,17 +124,18 @@ function LinkControlTransforms( { clientId } ) {
 }
 
 export function LinkUI( props ) {
+	const { label, url, opensInNewTab, type, kind } = props.link;
 	const link = {
-		url: props.link.url,
-		opensInNewTab: props.link.opensInNewTab,
-		title: props.link.label,
+		url,
+		opensInNewTab,
+		title: label && stripHTML( label ),
 	};
 
 	return (
 		<Popover
 			placement="bottom"
-			onClose={ props?.onClose }
-			anchor={ props?.anchor }
+			onClose={ props.onClose }
+			anchor={ props.anchor }
 			shift
 		>
 			<LinkControl
@@ -142,20 +144,17 @@ export function LinkUI( props ) {
 				className={ props.className }
 				value={ link }
 				showInitialSuggestions={ true }
-				withCreateSuggestion={ props?.hasCreateSuggestion }
-				noDirectEntry={ !! props.link?.type }
-				noURLSuggestion={ !! props.link?.type }
-				suggestionsQuery={ getSuggestionsQuery(
-					props.link?.type,
-					props.link?.kind
-				) }
+				withCreateSuggestion={ props.hasCreateSuggestion }
+				noDirectEntry={ !! type }
+				noURLSuggestion={ !! type }
+				suggestionsQuery={ getSuggestionsQuery( type, kind ) }
 				onChange={ props.onChange }
 				onRemove={ props.onRemove }
 				renderControlBottom={
-					! props.link?.url
+					! url
 						? () => (
 								<LinkControlTransforms
-									clientId={ props?.clientId }
+									clientId={ props.clientId }
 								/>
 						  )
 						: null