+++ packages/edit-site/src/components/sidebar/index.js
@@ -1,0 +1,19 @@
+/**
+ * WordPress dependencies
+ */
+import { __experimentalVStack as VStack } from '@wordpress/components';
+
+/**
+ * Internal dependencies
+ */
+import SidebarNavigationRoot from '../sidebar-navigation-root';
+
+export function Sidebar() {
+	return (
+		<VStack alignment="top" spacing={ 10 }>
+			<div className="edit-site-sidebar__content">
+				<SidebarNavigationRoot />
+			</div>
+		</VStack>
+	);
+}